backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.4.2
packaging==26.0
pandas==3.0.1
patsy==1.0.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.22
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
starlette==0.52.1
statsmodels==0.14.6
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
vaderSentiment==3.3.2
openpyxl==3.1.2
google-generativeai==0.8.6
google-genai==1.12.1
python-dotenv==1.1.1
